add user disconnection
